cli: Split out GetNotaryRepository and associated functions
Split these into cli/trust so that other commands can make use of them.

(cherry picked from commit 4b8c79f25ee00ca5dfe22271c166938009bda976)
